requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyMySQL==1.1.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
setuptools==71.0.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
